Removed "high level" custom object importers
The overhead is surprising and the value proposition over directly interacting with the tokenizer is dubious.  It's also possible to re-implement this feature by first parsing a JsonValue in the importer function before constructing a value of the new type.  Probably best to make the default path also the fast path.